+++ src/widgets/DataProgressBar.cs
@@ -159,7 +159,7 @@ namespace Gtk.DataBindings
 		{
 			adaptor.DataChanged = false;
 			object val = adaptor.Value;
-			double fval = (double) System.Convert.ToDouble(val); 
+			double fval = System.Convert.ToDouble(val); 
 			if (fval != Fraction)
 				Fraction = fval;
 		}
